Move logging to specific component
