+++ pkg/component/plutono/logging.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package monitoring
+package plutono
 
 import (
 	"fmt"
@@ -27,12 +27,7 @@ import (
 	"github.com/gardener/gardener/pkg/component"
 )
 
-var (
-	alermanagerName = "alertmanager"
-	prometheusName  = "prometheus"
-)
-
-// CentralLoggingConfiguration returns a fluent-bit parser and filter for the monitoring logs.
+// CentralLoggingConfiguration returns a fluent-bit parser and filter for the plutono logs.
 func CentralLoggingConfiguration() (component.CentralLoggingConfig, error) {
 	return component.CentralLoggingConfig{Filters: generateClusterFilters(), Parsers: generateClusterParsers()}, nil
 }
@@ -41,52 +36,16 @@ func generateClusterFilters() []*fluentbitv1alpha2.ClusterFilter {
 	return []*fluentbitv1alpha2.ClusterFilter{
 		{
 			ObjectMeta: metav1.ObjectMeta{
-				Name:   alermanagerName,
+				Name:   name,
 				Labels: map[string]string{v1beta1constants.LabelKeyCustomLoggingResource: v1beta1constants.LabelValueCustomLoggingResource},
 			},
 			Spec: fluentbitv1alpha2.FilterSpec{
-				Match: fmt.Sprintf("kubernetes.*%s*%s*", alermanagerName, alermanagerName),
+				Match: fmt.Sprintf("kubernetes.*%s*%s*", name, name),
 				FilterItems: []fluentbitv1alpha2.FilterItem{
 					{
 						Parser: &fluentbitv1alpha2filter.Parser{
 							KeyName:     "log",
-							Parser:      alermanagerName + "-parser",
-							ReserveData: pointer.Bool(true),
-						},
-					},
-				},
-			},
-		},
-		{
-			ObjectMeta: metav1.ObjectMeta{
-				Name:   fmt.Sprintf("%s--%s", prometheusName, prometheusName),
-				Labels: map[string]string{v1beta1constants.LabelKeyCustomLoggingResource: v1beta1constants.LabelValueCustomLoggingResource},
-			},
-			Spec: fluentbitv1alpha2.FilterSpec{
-				Match: fmt.Sprintf("kubernetes.*%s*%s*", prometheusName, prometheusName),
-				FilterItems: []fluentbitv1alpha2.FilterItem{
-					{
-						Parser: &fluentbitv1alpha2filter.Parser{
-							KeyName:     "log",
-							Parser:      prometheusName + "-parser",
-							ReserveData: pointer.Bool(true),
-						},
-					},
-				},
-			},
-		},
-		{
-			ObjectMeta: metav1.ObjectMeta{
-				Name:   prometheusName + "--blackbox-exporter",
-				Labels: map[string]string{v1beta1constants.LabelKeyCustomLoggingResource: v1beta1constants.LabelValueCustomLoggingResource},
-			},
-			Spec: fluentbitv1alpha2.FilterSpec{
-				Match: fmt.Sprintf("kubernetes.*%s*blackbox-exporter*", prometheusName),
-				FilterItems: []fluentbitv1alpha2.FilterItem{
-					{
-						Parser: &fluentbitv1alpha2filter.Parser{
-							KeyName:     "log",
-							Parser:      prometheusName + "-parser",
+							Parser:      name + "-parser",
 							ReserveData: pointer.Bool(true),
 						},
 					},
@@ -100,27 +59,14 @@ func generateClusterParsers() []*fluentbitv1alpha2.ClusterParser {
 	return []*fluentbitv1alpha2.ClusterParser{
 		{
 			ObjectMeta: metav1.ObjectMeta{
-				Name:   alermanagerName + "-parser",
+				Name:   name + "-parser",
 				Labels: map[string]string{v1beta1constants.LabelKeyCustomLoggingResource: v1beta1constants.LabelValueCustomLoggingResource},
 			},
 			Spec: fluentbitv1alpha2.ParserSpec{
 				Regex: &fluentbitv1alpha2parser.Regex{
-					Regex:      "^level=(?<severity>\\w+)\\s+ts=(?<time>\\d{4}-\\d{2}-\\d{2}[Tt].*[zZ])\\s+caller=(?<source>[^\\s]*+)\\s+(?<log>.*)",
+					Regex:      " ^t=(?<time>\\d{4}-\\d{2}-\\d{2}T[^ ]*)\\s+lvl=(?<severity>\\w+)\\smsg=\"(?<log>.*)\"\\s+logger=(?<source>.*)",
 					TimeKey:    "time",
-					TimeFormat: "%Y-%m-%dT%H:%M:%S.%L",
-				},
-			},
-		},
-		{
-			ObjectMeta: metav1.ObjectMeta{
-				Name:   prometheusName + "-parser",
-				Labels: map[string]string{v1beta1constants.LabelKeyCustomLoggingResource: v1beta1constants.LabelValueCustomLoggingResource},
-			},
-			Spec: fluentbitv1alpha2.ParserSpec{
-				Regex: &fluentbitv1alpha2parser.Regex{
-					Regex:      "^ts=(?<time>\\d{4}-\\d{2}-\\d{2}[Tt]{1}\\d{2}:\\d{2}:\\d{2}\\.\\d+\\S+)\\s+caller=(?<source>.+?)\\s+level=(?<severity>\\w+)\\s+(?<log>.*)$",
-					TimeKey:    "time",
-					TimeFormat: "%Y-%m-%dT%H:%M:%S.%L%z",
+					TimeFormat: "%Y-%m-%dT%H:%M:%S%z",
 				},
 			},
 		},
